services/insight-generator/requirements.txt
psycopg2-binary==2.9.9
schedule==1.2.2
python-dotenv==1.0.1
